requirements.txt
DeepFace
psycopg2
python-dotenv
